+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 #### 3.01.2019
 
+* [General] Converted VideoListAdapter to Kotlin
 * [General] Converted FavouriteListAdapter to Kotlin
 
 #### 2.01.2019